% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{FizzBuzz}
\usage{
FizzBuzz(n)
}
\arguments{
\item{n}{is a positive integer}
}
\value{
Fizz, Buzz, or FizzBuzz
}
\description{
This function counts as high as possible and replaces numbers divisible by 3
with Fizz, numbers divisible by 5 with Buzz, and numbers divisible by both
with FizzBuzz
}
