% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{Pulliam Flagstaff Weather Data}
\format{
A data frame with columns:
\describe{
  \item{DATE}{The date of the observation}
  \item{PRCP}{Precipitation (inches)}
  \item{SNOW}{Snowfall (inches)}
  \item{TMAX}{Maximum temperature (F)}
  \item{TMIN}{Minimum temperature (F)}
}
}
\source{
NOAA
}
\usage{
Flagstaff_Weather
}
\description{
A dataset containing daily weather observations from Flagstaff Pulliam Airport.
}
\keyword{datasets}
